añade Cajas a Trips

+++ database/migrations/tenant/2019_04_16_223230_add_truck_tract_relationship_to_trips_table.php
@@ -14,7 +14,7 @@ class AddTruckTractRelationshipToTripsTable extends Migration
     public function up()
     {
         Schema::table('trips', function (Blueprint $table) {
-            $table->integer('truck_tract_id')->unsigned();
+            $table->bigInteger('truck_tract_id')->unsigned();
         });
 
         Schema::table('trips', function (Blueprint $table){
